pyproject.toml
[project]
name = "flowerpot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
]
